Create Day08_circular
This commit includes the R script for Day 8 of the 30DayChartChallenge. The script uses data from the 2015-2016 Premier League season to analyze fouls committed both at home and away by each team. It employs advanced grouping and visualization techniques such as bar charts and polar coordinates to highlight differences in team behavior in different settings. The script relies on the 'season-1516.csv' dataset previously uploaded to the repository.